+++ app/users/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <div>
+      <div>マイページ作成予定</div>
+    </div>
+  );
+}
