+++ types/cartItem .ts
@@ -1,9 +1,8 @@
 export type CartItem = {
   id: string;
   name: string;
-  description: string;
   price: number;
   image: string;
   quantity: number;
-  grind: number;
+  grind: string;
 };
